requirements.txt
matplotlib==3.10.7
numpy==2.3.5
onnx==1.16.0
onnxruntime==1.23.2
onnxruntime_gpu==1.23.2
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
Pillow==12.0.0
tensorflow==2.20.0
